+++ server/camcops_server/cc_modules/cc_taskreports.py
@@ -28,6 +28,7 @@ camcops_server/cc_modules/cc_taskreports.py
 
 """
 
+from collections import Counter
 from typing import Any, List, Sequence, Type, TYPE_CHECKING
 
 from cardinal_pythonlib.classes import classproperty
@@ -46,6 +47,7 @@ from camcops_server.cc_modules.cc_pyramid import ViewParam
 from camcops_server.cc_modules.cc_report import Report, PlainReportType
 from camcops_server.cc_modules.cc_task import Task
 from camcops_server.cc_modules.cc_taskindex import TaskIndexEntry
+from camcops_server.cc_modules.cc_user import User
 
 if TYPE_CHECKING:
     from camcops_server.cc_modules.cc_request import CamcopsRequest
@@ -72,7 +74,10 @@ class TaskCountReport(Report):
     # noinspection PyMethodParameters
     @classproperty
     def title(cls) -> str:
-        return "(Server) Count current task instances, by creation date"
+        return (
+            "(Server) Count current task instances, by creation month and "
+            "task type"
+        )
 
     # noinspection PyMethodParameters
     @classproperty
@@ -100,53 +105,149 @@ class TaskCountReport(Report):
             # noinspection PyUnresolvedReferences
             query = (
                 select([
-                    TaskIndexEntry.task_table_name.label("task"),
                     extract_year(TaskIndexEntry.when_created_utc).label("year"),  # noqa
                     extract_month(TaskIndexEntry.when_created_utc).label("month"),  # noqa
+                    TaskIndexEntry.task_table_name.label("task"),
                     func.count().label("num_tasks_added"),
                 ])
                 .select_from(TaskIndexEntry.__table__)
-                .group_by("task", "year", "month")
-                .order_by("task", desc("year"), desc("month"))
+                .group_by("year", "month", "task")
+                .order_by(desc("year"), desc("month"), "task")
+                # ... http://docs.sqlalchemy.org/en/latest/core/tutorial.html#ordering-or-grouping-by-a-label  # noqa
             )
             if not superuser:
                 # Restrict to accessible groups
                 # noinspection PyProtectedMember
                 query = query.where(TaskIndexEntry.group_id.in_(group_ids))
             rows, colnames = get_rows_fieldnames_from_query(dbsession, query)
-            final_rows.extend(rows)
+            final_rows = rows
         else:
             classes = Task.all_subclasses_by_tablename()
             for cls in classes:
-                # noinspection PyProtectedMember
-                select_fields = [
-                    literal(cls.__tablename__).label("task"),
-                    # func.year() is specific to some DBs, e.g. MySQL
-                    # so is func.extract();
-                    # http://modern-sql.com/feature/extract
-                    extract_year(isotzdatetime_to_utcdatetime(
-                        cls.when_created)).label("year"),
-                    extract_month(isotzdatetime_to_utcdatetime(
-                        cls.when_created)).label("month"),
-                    func.count().label("num_tasks_added"),
-                ]
                 # noinspection PyUnresolvedReferences
-                select_from = cls.__table__
-                # noinspection PyProtectedMember
-                wheres = [cls._current == True]  # nopep8
+                query = (
+                    select([
+                        # func.year() is specific to some DBs, e.g. MySQL
+                        # so is func.extract();
+                        # http://modern-sql.com/feature/extract
+                        extract_year(isotzdatetime_to_utcdatetime(
+                            cls.when_created)).label("year"),
+                        extract_month(isotzdatetime_to_utcdatetime(
+                            cls.when_created)).label("month"),
+                        literal(cls.__tablename__).label("task"),
+                        func.count().label("num_tasks_added"),
+                    ])
+                    .select_from(cls.__table__)
+                    .where(cls._current == True)  # nopep8
+                    .group_by("year", "month")
+                )
                 if not superuser:
                     # Restrict to accessible groups
                     # noinspection PyProtectedMember
-                    wheres.append(cls._group_id.in_(group_ids))
-                group_by = ["year", "month"]
-                order_by = [desc("year"), desc("month")]
-                # ... http://docs.sqlalchemy.org/en/latest/core/tutorial.html#ordering-or-grouping-by-a-label  # noqa
-                query = select(select_fields) \
-                    .select_from(select_from) \
-                    .where(and_(*wheres)) \
-                    .group_by(*group_by) \
-                    .order_by(*order_by)
+                    query = query.where(cls._group_id.in_(group_ids))
                 rows, colnames = get_rows_fieldnames_from_query(
                     dbsession, query)
                 final_rows.extend(rows)
+            final_rows.sort(key=lambda r: (-r[0], -r[1], r[2]))
+            # ... sort by: year (descending), month (descending), task
+        return PlainReportType(rows=final_rows, column_names=colnames)
+
+
+class TaskCountByUserReport(Report):
+    """
+    Report to count task instances.
+    """
+
+    # noinspection PyMethodParameters
+    @classproperty
+    def report_id(cls) -> str:
+        return "taskcount_by_user"
+
+    # noinspection PyMethodParameters
+    @classproperty
+    def title(cls) -> str:
+        return (
+            "(Server) Count current task instances, by creation month and user"
+        )
+
+    # noinspection PyMethodParameters
+    @classproperty
+    def superuser_only(cls) -> bool:
+        return False
+
+    @staticmethod
+    def get_paramform_schema_class() -> Type["ReportParamSchema"]:
+        return TaskCountReportSchema
+
+    @classmethod
+    def get_specific_http_query_keys(cls) -> List[str]:
+        return [
+            ViewParam.VIA_INDEX
+        ]
+
+    def get_rows_colnames(self, req: "CamcopsRequest") -> PlainReportType:
+        final_rows = []  # type: List[Sequence[Sequence[Any]]]
+        colnames = []  # type: List[str]
+        dbsession = req.dbsession
+        group_ids = req.user.ids_of_groups_user_may_report_on
+        superuser = req.user.superuser
+        via_index = req.get_bool_param(ViewParam.VIA_INDEX, True)
+        if via_index:
+            # noinspection PyUnresolvedReferences
+            query = (
+                select([
+                    extract_year(TaskIndexEntry.when_created_utc).label("year"),  # noqa
+                    extract_month(TaskIndexEntry.when_created_utc).label("month"),  # noqa
+                    User.username.label("adding_user_name"),
+                    func.count().label("num_tasks_added"),
+                ])
+                .select_from(TaskIndexEntry.__table__)
+                .select_from(User.__table__)
+                .where(TaskIndexEntry.adding_user_id == User.id)
+                .group_by("year", "month", "adding_user_name")
+                .order_by(desc("year"), desc("month"), "adding_user_name")
+            )
+            if not superuser:
+                # Restrict to accessible groups
+                # noinspection PyProtectedMember
+                query = query.where(TaskIndexEntry.group_id.in_(group_ids))
+            rows, colnames = get_rows_fieldnames_from_query(dbsession, query)
+            final_rows = rows
+        else:
+            classes = Task.all_subclasses_by_tablename()
+            counter = Counter()
+            for cls in classes:
+                # noinspection PyUnresolvedReferences
+                query = (
+                    select([
+                        # func.year() is specific to some DBs, e.g. MySQL
+                        # so is func.extract();
+                        # http://modern-sql.com/feature/extract
+                        extract_year(isotzdatetime_to_utcdatetime(
+                            cls.when_created)).label("year"),
+                        extract_month(isotzdatetime_to_utcdatetime(
+                            cls.when_created)).label("month"),
+                        User.username.label("adding_user_name"),
+                        func.count().label("num_tasks_added"),
+                    ])
+                    .select_from(cls.__table__)
+                    .select_from(User.__table__)
+                    .where(and_(cls._current == True,  # nopep8
+                                cls._adding_user_id == User.id))
+                    .group_by("year", "month", "adding_user_name")
+                )
+                if not superuser:
+                    # Restrict to accessible groups
+                    # noinspection PyProtectedMember
+                    query = query.where(cls._group_id.in_(group_ids))
+                rows, colnames = get_rows_fieldnames_from_query(
+                    dbsession, query)
+                for row in rows:
+                    year, month, username = row[0], row[1], row[2]
+                    count = row[3]
+                    counter.update({(year, month, username): count})
+            for (year, month, username), total in counter.items():
+                final_rows.append([year, month, username, total])
+            final_rows.sort(key=lambda r: (-r[0], -r[1], r[2]))
+            # ... sort by: year (descending), month (descending), username
         return PlainReportType(rows=final_rows, column_names=colnames)